Enemies spawn closer to "sweet spot" of fun game
Ensure Enem(y/ies) spawn within a certain Goldie Locks range of the
player's starting position.

In pursuance of this, add area method to rect.

While here, fix an instance of unnecessary static casting.

+++ src/classes/Actor.cpp
@@ -191,9 +191,9 @@ void Actor::setPos(vec2ui newPos) {
 };
 
 void Actor::setPosRand() {
-    int_fast8_t randx = (rand() % (game_area.right() -1) + 1);
-    int_fast8_t randy = (rand() % (game_area.bot()   -1) + 1);
-    setPos({ static_cast<uint_fast8_t>(randx), static_cast<uint_fast8_t>(randy) });
+    uint_fast8_t randx = (rand() % (game_area.right() -1) + 1);
+    uint_fast8_t randy = (rand() % (game_area.bot()   -1) + 1);
+    setPos({ randx, randy });
 };
 
 void Actor::setColo(chtype newColo) {
